active-response: restore network compatibility wrapper
