+++ src/security/active_response_impl.rs
@@ -1,0 +1,2449 @@
+//! Active response: reversible, auditable intervention actions.
+//!
+//! Phase 11 starts with practical controls for blocking traffic, killing a
+//! live socket, suspending a suspicious process, blocking a suspicious domain,
+//! and isolating the machine. The module persists a tiny state file so rules
+//! can be reconciled and the UI can reflect the current status after restarts.
+
+use crate::security::firewall::{self, FirewallBackend};
+use crate::{audit, quarantine, types::ConnInfo};
+use serde::{Deserialize, Serialize};
+use serde_json::json;
+use sha2::{Digest, Sha256};
+use std::net::{IpAddr, SocketAddr, TcpStream};
+use std::path::{Path, PathBuf};
+use std::sync::{Mutex, OnceLock, RwLock};
+use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
+
+const STATE_FILE: &str = "vigil-active-response.json";
+const BLOCK_RULE_PREFIX: &str = "Vigil Block";
+const PROCESS_BLOCK_RULE_PREFIX: &str = "Vigil Proc Block";
+const DOMAIN_MARKER_PREFIX: &str = "# Vigil Domain Block";
+const ISOLATE_RULE_IN: &str = "Vigil Isolate In";
+const ISOLATE_RULE_OUT: &str = "Vigil Isolate Out";
+const ISOLATION_MAX_SECS: u64 = 60 * 60;
+const ISOLATION_ACTIVATION_GRACE_SECS: u64 = 20;
+const PROCESS_RULE_PREFIX_LEN: usize = 48;
+const PROCESS_RULE_FINGERPRINT_LEN: usize = 16;
+const QUERY_STATE_CACHE_TTL: Duration = Duration::from_millis(250);
+
+fn backend() -> &'static dyn FirewallBackend {
+    firewall::get_backend()
+}
+
+#[path = "active_response_platform.rs"]
+mod platform;
+
+#[derive(Debug, Clone, Copy, Default, Serialize, Deserialize)]
+pub struct Status {
+    pub blocked_rules: usize,
+    pub blocked_processes: usize,
+    pub blocked_domains: usize,
+    pub suspended_processes: usize,
+    pub frozen_autoruns: bool,
+    pub isolated: bool,
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct InspectorSnapshot {
+    pub status: Status,
+    pub remote_target: Option<String>,
+    pub remote_blocked: bool,
+    pub remote_remaining: Option<Duration>,
+    pub domain_target: Option<String>,
+    pub domain_blocked: bool,
+    pub process_blocked: bool,
+    pub process_remaining: Option<Duration>,
+    pub process_suspended: bool,
+    pub connection_kill_enabled: bool,
+    pub quarantine_ready: bool,
+    pub suspend_enabled: bool,
+    pub firewall_modifiable: bool,
+    pub network_isolation_modifiable: bool,
+    pub domain_modifiable: bool,
+}
+
+#[derive(Debug, Clone, Default, Serialize)]
+pub struct FirewallRuleList {
+    pub blocked_ips: Vec<BlockedIpEntry>,
+    pub blocked_processes: Vec<BlockedProcessEntry>,
+    pub blocked_domains: Vec<BlockedDomainEntry>,
+    pub suspended_processes: Vec<SuspendedProcessEntry>,
+    pub profiles: Vec<FirewallProfileEntry>,
+    pub isolated: bool,
+    pub isolation_remaining_secs: Option<u64>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+pub struct BlockedIpEntry {
+    pub target: String,
+    pub rule_name: String,
+    pub expires_at_unix: Option<u64>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+pub struct BlockedProcessEntry {
+    pub pid: u32,
+    pub path: String,
+    pub inbound_rule_name: String,
+    pub outbound_rule_name: String,
+    pub expires_at_unix: Option<u64>,
+}
+
+#[derive(Debug, Clone, Serialize)]
+pub struct BlockedDomainEntry {
+    pub domain: String,
+}
+
+#[derive(Debug, Clone, Serialize)]
+pub struct SuspendedProcessEntry {
+    pub pid: u32,
+    pub path: String,
+    pub proc_name: String,
+    pub suspended_at_unix: u64,
+}
+
+#[derive(Debug, Clone, Serialize)]
+pub struct FirewallProfileEntry {
+    pub name: String,
+    pub enabled: bool,
+    pub inbound_action: String,
+    pub outbound_action: String,
+}
+
+/// Load the current firewall rule list for the UI tab.
+/// Uses the same caching as `status()` — safe to call every frame.
+pub fn list_rules() -> FirewallRuleList {
+    let mut result = FirewallRuleList::default();
+    let state = match load_state_for_query("list_rules") {
+        Some(s) => s,
+        None => return result,
+    };
+    let now = unix_now();
+    result.blocked_ips = state
+        .blocked
+        .into_iter()
+        .map(|b| BlockedIpEntry {
+            target: b.target,
+            rule_name: b.rule_name,
+            expires_at_unix: b.expires_at_unix,
+        })
+        .collect();
+    result.blocked_processes = state
+        .blocked_processes
+        .into_iter()
+        .map(|b| BlockedProcessEntry {
+            pid: b.pid,
+            path: b.path,
+            inbound_rule_name: b.inbound_rule_name,
+            outbound_rule_name: b.outbound_rule_name,
+            expires_at_unix: b.expires_at_unix,
+        })
+        .collect();
+    result.blocked_domains = state
+        .blocked_domains
+        .into_iter()
+        .map(|b| BlockedDomainEntry { domain: b.domain })
+        .collect();
+    result.suspended_processes = state
+        .suspended_processes
+        .into_iter()
+        .map(|s| SuspendedProcessEntry {
+            pid: s.pid,
+            path: s.path,
+            proc_name: s.proc_name,
+            suspended_at_unix: s.suspended_at_unix,
+        })
+        .collect();
+    result.isolated = state.isolated;
+    if let Some(expires) = state.isolation_expires_unix {
+        result.isolation_remaining_secs = expires.checked_sub(now);
+    }
+    if let Some(snapshot) = state.firewall_snapshot {
+        result.profiles = snapshot
+            .profiles
+            .into_iter()
+            .map(|p| FirewallProfileEntry {
+                name: p.name,
+                enabled: p.enabled,
+                inbound_action: p.inbound_action,
+                outbound_action: p.outbound_action,
+            })
+            .collect();
+    }
+    result
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+struct State {
+    blocked: Vec<BlockedTarget>,
+    #[serde(default)]
+    blocked_processes: Vec<BlockedProcess>,
+    #[serde(default)]
+    blocked_domains: Vec<BlockedDomain>,
+    #[serde(default)]
+    suspended_processes: Vec<SuspendedProcess>,
+    #[serde(default)]
+    autorun_snapshot: Option<AutorunSnapshot>,
+    #[serde(default)]
+    firewall_snapshot: Option<FirewallSnapshot>,
+    #[serde(default)]
+    network_snapshot: Option<NetworkSnapshot>,
+    #[serde(default)]
+    isolation_started_unix: Option<u64>,
+    #[serde(default)]
+    isolation_expires_unix: Option<u64>,
+    isolated: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct BlockedTarget {
+    target: String,
+    rule_name: String,
+    expires_at_unix: Option<u64>,
+}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct BlockedProcess {
+    #[serde(default)]
+    pid: u32,
+    path: String,
+    inbound_rule_name: String,
+    outbound_rule_name: String,
+    expires_at_unix: Option<u64>,
+}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct BlockedDomain {
+    domain: String,
+    marker: String,
+}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct SuspendedProcess {
+    pid: u32,
+    path: String,
+    proc_name: String,
+    suspended_at_unix: u64,
+}
+#[derive(Debug, Clone, Serialize, Deserialize)]
+struct AutorunSnapshot {
+    captured_at_unix: u64,
+    entries: Vec<AutorunEntry>,
+}
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+struct AutorunEntry {
+    hive: String,
+    key_path: String,
+    value_name: String,
+    value_data: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+struct FirewallProfileState {
+    name: String,
+    enabled: bool,
+    inbound_action: String,
+    outbound_action: String,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+struct FirewallSnapshot {
+    profiles: Vec<FirewallProfileState>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+struct NetworkAdapterState {
+    name: String,
+    #[serde(default)]
+    is_wireless: bool,
+    #[serde(default)]
+    wifi_profile: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+struct NetworkSnapshot {
+    adapters: Vec<NetworkAdapterState>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[allow(dead_code)]
+struct TcpSessionState {
+    local_address: String,
+    local_port: u16,
+    remote_address: String,
+    remote_port: u16,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum SocketKillError {
+    UnsupportedStatus(String),
+    InvalidLocalAddr(String),
+    InvalidRemoteAddr(String),
+    UnsupportedAddressFamily,
+    PlatformUnsupported,
+    PermissionDenied,
+    OsError(String),
+}
+impl std::fmt::Display for SocketKillError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Self::UnsupportedStatus(status) => write!(f, "cannot kill a socket in {status} state"),
+            Self::InvalidLocalAddr(addr) => write!(f, "invalid local address: {addr}"),
+            Self::InvalidRemoteAddr(addr) => write!(f, "invalid remote address: {addr}"),
+            Self::UnsupportedAddressFamily => write!(
+                f,
+                "live socket kill currently supports IPv4 TCP only on Windows"
+            ),
+            Self::PlatformUnsupported => {
+                write!(f, "socket kill is currently only implemented on Windows")
+            }
+            Self::PermissionDenied => write!(
+                f,
+                "administrator privileges are required to kill a TCP connection"
+            ),
+            Self::OsError(msg) => write!(f, "OS error: {msg}"),
+        }
+    }
+}
+impl std::error::Error for SocketKillError {}
+#[derive(Debug, Clone, PartialEq, Eq)]
+struct SocketKillTarget {
+    local: SocketAddr,
+    remote: SocketAddr,
+}
+
+#[allow(dead_code)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum DurationPreset {
+    OneHour,
+    OneDay,
+    Permanent,
+}
+impl DurationPreset {
+    fn ttl(self) -> Option<Duration> {
+        match self {
+            Self::OneHour => Some(Duration::from_secs(60 * 60)),
+            Self::OneDay => Some(Duration::from_secs(60 * 60 * 24)),
+            Self::Permanent => None,
+        }
+    }
+
+    pub fn label(self) -> &'static str {
+        match self {
+            Self::OneHour => "1h",
+            Self::OneDay => "24h",
+            Self::Permanent => "Permanent",
+        }
+    }
+}
+
+pub fn status() -> Status {
+    let state = match load_state() {
+        Ok(state) => state,
+        Err(err) => {
+            note_state_load_error_once("status", &err);
+            return Status {
+                isolated: platform::isolation_controls_active(&State::default()).unwrap_or(false),
+                ..Status::default()
+            };
+        }
+    };
+    let now = unix_now();
+    Status {
+        blocked_rules: state.blocked.len() + state.blocked_processes.len(),
+        blocked_processes: state.blocked_processes.len(),
+        blocked_domains: state.blocked_domains.len(),
+        suspended_processes: state.suspended_processes.len(),
+        frozen_autoruns: state.autorun_snapshot.is_some(),
+        // Reflect effective containment state, not just persisted intent.
+        isolated: isolation_effective_now(&state, now),
+    }
+}
+pub fn has_frozen_autoruns() -> bool {
+    load_state_for_query("has_frozen_autoruns")
+        .and_then(|s| s.autorun_snapshot)
+        .is_some()
+}
+pub fn can_modify_firewall() -> bool {
+    platform::is_supported() && platform::is_elevated()
+}
+pub fn can_isolate_network() -> bool {
+    platform::supports_isolation() && platform::is_elevated()
+}
+pub fn can_kill_connection(conn: &ConnInfo) -> bool {
+    if !platform::is_supported() || !platform::is_elevated() {
+        return false;
+    }
+    socket_kill_target(conn).is_ok()
+}
+pub fn can_suspend_process(pid: u32) -> bool {
+    platform::is_supported() && platform::is_elevated() && pid != 0
+}
+pub fn can_block_domain() -> bool {
+    platform::is_supported() && platform::is_elevated()
+}
+pub fn can_apply_quarantine_profile(pid: u32, path: &str) -> bool {
+    platform::is_supported() && platform::is_elevated() && (pid != 0 || !path.trim().is_empty())
+}
+
+pub fn inspector_snapshot(
+    pid: u32,
+    path: &str,
+    selected_connection: Option<&ConnInfo>,
+) -> InspectorSnapshot {
+    let state = load_state_for_query("inspector_snapshot");
+    let state_ref = state.as_ref();
+    let now = unix_now();
+    let status = state_ref.map_or_else(
+        || Status {
+            isolated: platform::isolation_controls_active(&State::default()).unwrap_or(false),
+            ..Status::default()
+        },
+        |state| Status {
+            blocked_rules: state.blocked.len() + state.blocked_processes.len(),
+            blocked_processes: state.blocked_processes.len(),
+            blocked_domains: state.blocked_domains.len(),
+            suspended_processes: state.suspended_processes.len(),
+            frozen_autoruns: state.autorun_snapshot.is_some(),
+            isolated: isolation_effective_now(state, now),
+        },
+    );
+    let firewall_modifiable = can_modify_firewall();
+    let network_isolation_modifiable = can_isolate_network();
+    let domain_modifiable = can_block_domain();
+    let suspend_enabled = can_suspend_process(pid);
+    let quarantine_ready = can_apply_quarantine_profile(pid, path);
+    let (remote_target, remote_blocked, remote_remaining, domain_target, domain_blocked) =
+        if let Some(conn) = selected_connection {
+            let remote_target = extract_remote_target(&conn.remote_addr);
+            let domain_target = extract_domain_target(conn);
+            let remote_blocked = remote_target.as_deref().is_some_and(|target| {
+                state_ref
+                    .is_some_and(|state| state.blocked.iter().any(|rule| rule.target == target))
+            });
+            let remote_remaining = remote_target.as_deref().and_then(|target| {
+                state_ref.and_then(|state| {
+                    state
+                        .blocked
+                        .iter()
+                        .find(|rule| rule.target == target)
+                        .and_then(|rule| rule.expires_at_unix)
+                        .and_then(|expires_at_unix| expires_at_unix.checked_sub(now))
+                        .map(Duration::from_secs)
+                })
+            });
+            let domain_blocked = domain_target.as_deref().is_some_and(|domain| {
+                state_ref.is_some_and(|state| {
+                    state
+                        .blocked_domains
+                        .iter()
+                        .any(|entry| entry.domain == domain)
+                })
+            });
+            (
+                remote_target,
+                remote_blocked,
+                remote_remaining,
+                domain_target,
+                domain_blocked,
+            )
+        } else {
+            (None, false, None, None, false)
+        };
+    let normalized_path = normalise_target(path).ok();
+    let process_blocked = state_ref
+        .zip(normalized_path.as_ref())
+        .is_some_and(|(state, path)| {
+            state
+                .blocked_processes
+                .iter()
+                .any(|rule| process_block_matches(rule, path))
+        });
+    let process_remaining = state_ref.and_then(|state| {
+        let path = normalise_target(path).ok()?;
+        state
+            .blocked_processes
+            .iter()
+            .find(|rule| process_block_matches(rule, &path))
+            .and_then(|rule| rule.expires_at_unix)
+            .and_then(|expires_at_unix| expires_at_unix.checked_sub(now))
+            .map(Duration::from_secs)
+    });
+    let process_suspended = state_ref.is_some_and(|state| {
+        state
+            .suspended_processes
+            .iter()
+            .any(|entry| suspended_process_matches(entry, pid, path))
+    });
+    let connection_kill_enabled = selected_connection.is_some_and(can_kill_connection);
+    InspectorSnapshot {
+        status,
+        remote_target,
+        remote_blocked,
+        remote_remaining,
+        domain_target,
+        domain_blocked,
+        process_blocked,
+        process_remaining,
+        process_suspended,
+        connection_kill_enabled,
+        quarantine_ready,
+        suspend_enabled,
+        firewall_modifiable,
+        network_isolation_modifiable,
+        domain_modifiable,
+    }
+}
+
+pub fn freeze_autoruns() -> Result<String, String> {
+    ensure_modifiable()?;
+    let snapshot = platform::snapshot_autoruns()?;
+    let mut state = load_state()?;
+    let count = snapshot.entries.len();
+    state.autorun_snapshot = Some(snapshot);
+    save_state(&state)?;
+    audit::record("freeze_autoruns", "success", json!({ "entries": count }));
+    Ok(format!(
+        "Captured autorun baseline with {count} entr{}.",
+        if count == 1 { "y" } else { "ies" }
+    ))
+}
+pub fn revert_frozen_autoruns() -> Result<String, String> {
+    ensure_modifiable()?;
+    let mut state = load_state()?;
+    let Some(snapshot) = state.autorun_snapshot.clone() else {
+        return Err("No autorun baseline has been captured yet.".into());
+    };
+    let result = platform::revert_autorun_changes(&snapshot.entries)?;
+    state.autorun_snapshot = None;
+    save_state(&state)?;
+    audit::record(
+        "revert_frozen_autoruns",
+        "success",
+        json!({ "removed_additions": result.removed_additions, "restored_entries": result.restored_entries }),
+    );
+    Ok(format!(
+        "Reverted autorun changes: removed {} added entr{} and restored {} baseline entr{}.",
+        result.removed_additions,
+        if result.removed_additions == 1 {
+            "y"
+        } else {
+            "ies"
+        },
+        result.restored_entries,
+        if result.restored_entries == 1 {
+            "y"
+        } else {
+            "ies"
+        }
+    ))
+}
+pub fn kill_connection(conn: &ConnInfo) -> Result<String, SocketKillError> {
+    if !platform::is_supported() {
+        return Err(SocketKillError::PlatformUnsupported);
+    }
+    if !platform::is_elevated() {
+        return Err(SocketKillError::PermissionDenied);
+    }
+    let target = socket_kill_target(conn)?;
+    backend()
+        .kill_tcp_connection(&target.local.to_string(), &target.remote.to_string())
+        .map_err(SocketKillError::OsError)?;
+    let message = format!(
+        "Killed TCP connection {} -> {}.",
+        target.local, target.remote
+    );
+    audit::record(
+        "kill_connection",
+        "success",
+        json!({ "local_addr": target.local.to_string(), "remote_addr": target.remote.to_string(), "pid": conn.pid, "proc_name": conn.proc_name }),
+    );
+    Ok(message)
+}
+pub fn suspend_process(pid: u32, path: &str, proc_name: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    if pid == 0 {
+        return Err("Cannot suspend PID 0.".into());
+    }
+    platform::suspend_process(pid)?;
+    let path = path.trim().to_string();
+    let proc_name = proc_name.trim().to_string();
+    let mut state = load_state()?;
+    state
+        .suspended_processes
+        .retain(|entry| !suspended_process_matches(entry, pid, &path));
+    state.suspended_processes.push(SuspendedProcess {
+        pid,
+        path: path.clone(),
+        proc_name: proc_name.clone(),
+        suspended_at_unix: unix_now(),
+    });
+    save_state(&state)?;
+    let message = if path.is_empty() {
+        format!("Suspended PID {pid}.")
+    } else {
+        format!("Suspended PID {pid} ({path}).")
+    };
+    audit::record(
+        "suspend_process",
+        "success",
+        json!({ "pid": pid, "path": path, "proc_name": proc_name }),
+    );
+    Ok(message)
+}
+pub fn resume_process(pid: u32, path: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    if pid == 0 {
+        return Err("Cannot resume PID 0.".into());
+    }
+    platform::resume_process(pid)?;
+    let path = path.trim().to_string();
+    let mut state = load_state()?;
+    let before = state.suspended_processes.len();
+    state
+        .suspended_processes
+        .retain(|entry| !suspended_process_matches(entry, pid, &path));
+    let removed = before.saturating_sub(state.suspended_processes.len());
+    save_state(&state)?;
+    let message = if path.is_empty() {
+        format!("Resumed PID {pid}.")
+    } else {
+        format!("Resumed PID {pid} ({path}).")
+    };
+    audit::record(
+        "resume_process",
+        if removed > 0 { "success" } else { "noop" },
+        json!({ "pid": pid, "path": path, "removed_entries": removed }),
+    );
+    Ok(message)
+}
+pub fn block_domain(domain: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    let domain = normalise_domain(domain)?;
+    let marker = domain_marker(&domain);
+    backend().add_domain_block(&domain, &marker)?;
+    let mut state = load_state()?;
+    state.blocked_domains.retain(|entry| entry.domain != domain);
+    state.blocked_domains.push(BlockedDomain {
+        domain: domain.clone(),
+        marker: marker.clone(),
+    });
+    save_state(&state)?;
+    audit::record(
+        "block_domain",
+        "success",
+        json!({ "domain": domain, "marker": marker }),
+    );
+    Ok(format!("Blocked domain {domain} via the local hosts file."))
+}
+pub fn unblock_domain(domain: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    let domain = normalise_domain(domain)?;
+    let marker = domain_marker(&domain);
+    backend().remove_domain_block(&domain, &marker)?;
+    let mut state = load_state()?;
+    let before = state.blocked_domains.len();
+    state.blocked_domains.retain(|entry| entry.domain != domain);
+    let removed = before.saturating_sub(state.blocked_domains.len());
+    save_state(&state)?;
+    audit::record(
+        "unblock_domain",
+        if removed > 0 { "success" } else { "noop" },
+        json!({ "domain": domain, "marker": marker, "removed_entries": removed }),
+    );
+    Ok(if removed > 0 {
+        format!("Removed the hosts-file block for {domain}.")
+    } else {
+        format!("No hosts-file block found for {domain}.")
+    })
+}
+
+pub fn apply_quarantine_profile(pid: u32, path: &str, proc_name: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    if pid == 0 && path.trim().is_empty() {
+        return Err("Quarantine profile requires a real PID or a known executable path.".into());
+    }
+    let mut applied = Vec::new();
+    let mut warnings = Vec::new();
+    match isolate_machine() {
+        Ok(_) => applied.push("network isolated".to_string()),
+        Err(err) => warnings.push(format!("network isolation failed: {err}")),
+    }
+    if !path.trim().is_empty() {
+        match block_process(pid, path, DurationPreset::Permanent) {
+            Ok(_) => applied.push("process traffic blocked".to_string()),
+            Err(err) => warnings.push(format!("process block failed: {err}")),
+        }
+    }
+    if pid != 0 {
+        match suspend_process(pid, path, proc_name) {
+            Ok(_) => applied.push("process suspended".to_string()),
+            Err(err) => warnings.push(format!("process suspend failed: {err}")),
+        }
+    }
+    let (extended_applied, extended_warnings) = quarantine::apply();
+    applied.extend(extended_applied);
+    warnings.extend(extended_warnings);
+    if applied.is_empty() {
+        return Err(if warnings.is_empty() {
+            "Quarantine profile could not apply any containment step.".into()
+        } else {
+            warnings.join("; ")
+        });
+    }
+    let message = if warnings.is_empty() {
+        format!("Applied quarantine profile: {}.", applied.join(", "))
+    } else {
+        format!(
+            "Applied quarantine profile: {}. Warnings: {}.",
+            applied.join(", "),
+            warnings.join("; ")
+        )
+    };
+    audit::record(
+        "quarantine_profile",
+        if warnings.is_empty() {
+            "success"
+        } else {
+            "partial"
+        },
+        json!({ "pid": pid, "path": path, "proc_name": proc_name, "applied": applied, "warnings": warnings }),
+    );
+    Ok(message)
+}
+
+pub fn clear_quarantine_profile(pid: u32, path: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    let mut cleared = Vec::new();
+    let mut warnings = Vec::new();
+    match restore_machine() {
+        Ok(_) => cleared.push("network restored".to_string()),
+        Err(err) => warnings.push(format!("network restore failed: {err}")),
+    }
+    if !path.trim().is_empty() {
+        match unblock_process(pid, path) {
+            Ok(_) => cleared.push("process block removed".to_string()),
+            Err(err) => warnings.push(format!("process unblock failed: {err}")),
+        }
+    }
+    if pid != 0 {
+        match resume_process(pid, path) {
+            Ok(_) => cleared.push("process resumed".to_string()),
+            Err(err) => warnings.push(format!("process resume failed: {err}")),
+        }
+    }
+    let (extended_cleared, extended_warnings) = quarantine::clear();
+    cleared.extend(extended_cleared);
+    warnings.extend(extended_warnings);
+    if cleared.is_empty() {
+        return Err(if warnings.is_empty() {
+            "Quarantine clear did not change any containment step.".into()
+        } else {
+            warnings.join("; ")
+        });
+    }
+    let message = if warnings.is_empty() {
+        format!("Cleared quarantine profile: {}.", cleared.join(", "))
+    } else {
+        format!(
+            "Cleared quarantine profile: {}. Warnings: {}.",
+            cleared.join(", "),
+            warnings.join("; ")
+        )
+    };
+    audit::record(
+        "clear_quarantine_profile",
+        if warnings.is_empty() {
+            "success"
+        } else {
+            "partial"
+        },
+        json!({ "pid": pid, "path": path, "cleared": cleared, "warnings": warnings }),
+    );
+    Ok(message)
+}
+
+/// Phase 19 P2: re-apply firewall rules that may have been lost on reboot.
+///
+/// On Linux, nftables/iptables rules are in-memory and disappear after a
+/// reboot. This function checks each blocked IP/process/domain from the
+/// persisted state against the live backend and re-applies any that are
+/// missing. On Windows, WFP filters survive reboots natively, so this is
+/// a no-op (the kernel persists Vigil's filters across restarts).
+#[cfg(not(windows))]
+fn is_domain_in_hosts(domain: &str, marker: &str) -> bool {
+    std::fs::read_to_string("/etc/hosts")
+        .map(|c| c.lines().any(|l| l.contains(domain) && l.contains(marker)))
+        .unwrap_or(false)
+}
+
+fn reconcile_firewall_rules() {
+    #[cfg(not(windows))]
+    {
+        let mut state = match load_state() {
+            Ok(s) => s,
+            Err(_) => return,
+        };
+        let now = unix_now();
+        let b = backend();
+        let mut reapplied = false;
+        let mut state_changed = false;
+        let mut deleted_count = 0usize;
+        let mut reapply_count = 0usize;
+
+        let blocked_before = state.blocked.len();
+        let proc_blocked_before = state.blocked_processes.len();
+
+        // Delete live rules for expired entries before removing from state.
+        // On a process restart (no reboot), iptables rules survive while
+        // Vigil was down — pruning state without deleting the live rule
+        // orphans it. Only remove from state when deletion succeeds.
+        let mut deleted_blocked: std::collections::HashSet<String> =
+            std::collections::HashSet::new();
+        let mut deleted_process_paths: std::collections::HashSet<String> =
+            std::collections::HashSet::new();
+
+        for rule in &state.blocked {
+            if rule.expires_at_unix.is_none_or(|deadline| deadline > now) {
+                continue;
+            }
+            if b.delete_rule(&rule.rule_name).is_ok() {
+                deleted_count += 1;
+                deleted_blocked.insert(rule.target.clone());
+            }
+        }
+        for rule in &state.blocked_processes {
+            if rule.expires_at_unix.is_none_or(|deadline| deadline > now) {
+                continue;
+            }
+            let out_ok = b.delete_rule(&rule.outbound_rule_name).is_ok();
+            let in_ok = b.delete_rule(&rule.inbound_rule_name).is_ok();
+            if out_ok && in_ok {
+                deleted_count += 2;
+                deleted_process_paths.insert(rule.path.clone());
+            }
+        }
+
+        state.blocked.retain(|rule| {
+            rule.expires_at_unix.is_none_or(|deadline| deadline > now)
+                || !deleted_blocked.contains(&rule.target)
+        });
+        state.blocked_processes.retain(|rule| {
+            rule.expires_at_unix.is_none_or(|deadline| deadline > now)
+                || !deleted_process_paths.contains(&rule.path)
+        });
+        if state.blocked.len() != blocked_before
+            || state.blocked_processes.len() != proc_blocked_before
+        {
+            state_changed = true;
+        }
+
+        // Re-apply IP blocks.
+        for rule in &state.blocked {
+            if b.rule_present(&rule.rule_name).unwrap_or(false) {
+                continue;
+            }
+            if b.add_block_rule(&rule.rule_name, &rule.target).is_ok() {
+                reapply_count += 1;
+                reapplied = true;
+            }
+        }
+
+        // Re-apply process blocks by looking up the current PID from
+        // the saved executable path.
+        if state.blocked_processes.iter().any(|bp| !bp.path.is_empty()) {
+            use sysinfo::{Pid, ProcessesToUpdate, System};
+            let mut sys = System::new();
+            sys.refresh_processes(ProcessesToUpdate::All, true);
+            for rule in &state.blocked_processes {
+                if rule.path.is_empty() {
+                    continue;
+                }
+                if b.rule_present(&rule.outbound_rule_name).unwrap_or(false)
+                    && b.rule_present(&rule.inbound_rule_name).unwrap_or(false)
+                {
+                    continue;
+                }
+                let live_pid = sys.processes().iter().find_map(|(pid, proc)| {
+                    let exe = proc.exe()?;
+                    if exe.to_string_lossy().as_ref() == rule.path.as_str() {
+                        Some(pid.as_u32())
+                    } else {
+                        None
+                    }
+                });
+                let Some(pid) = live_pid else { continue };
+                if !b.rule_present(&rule.outbound_rule_name).unwrap_or(false) {
+                    if b.add_block_program_rule(&rule.outbound_rule_name, pid, &rule.path, "out")
+                        .is_ok()
+                    {
+                        reapply_count += 1;
+                        reapplied = true;
+                    }
+                }
+                if !b.rule_present(&rule.inbound_rule_name).unwrap_or(false) {
+                    if b.add_block_program_rule(&rule.inbound_rule_name, pid, &rule.path, "in")
+                        .is_ok()
+                    {
+                        reapply_count += 1;
+                        reapplied = true;
+                    }
+                }
+            }
+        }
+
+        // Re-apply domain blocks that may have been lost from the hosts file.
+        for domain in &state.blocked_domains {
+            if is_domain_in_hosts(&domain.domain, &domain.marker) {
+                continue;
+            }
+            if b.add_domain_block(&domain.domain, &domain.marker).is_ok() {
+                reapply_count += 1;
+                reapplied = true;
+            }
+        }
+
+        if state.isolated {
+            if !b
+                .rule_present(crate::security::active_response::ISOLATE_RULE_IN)
+                .unwrap_or(false)
+                && b.apply_isolation("Vigil Isolate").is_ok()
+            {
+                reapply_count += 1;
+                reapplied = true;
+            }
+        }
+
+        if !state_changed && !reapplied {
+            return;
+        }
+        if state_changed {
+            if let Err(err) = save_state(&state) {
+                note_state_load_error_once("reconcile_firewall", &err);
+            }
+        }
+        if reapplied {
+            tracing::info!(
+                deleted = deleted_count,
+                reapplied = reapply_count,
+                "reconcile_firewall_rules: startup reconciliation complete"
+            );
+        }
+        backend().save_boot_config();
+    }
+    #[cfg(windows)]
+    {
+        // WFP filters persist in the kernel across reboots.
+    }
+}
+
+pub fn reconcile() {
+    if !platform::is_elevated() {
+        return;
+    }
+    if !platform::is_supported() && !platform::supports_isolation() {
+        return;
+    }
+    let Ok(mut state) = load_state() else {
+        note_state_load_error_once(
+            "reconcile",
+            "failed to load protected active-response state",
+        );
+        return;
+    };
+    let now = unix_now();
+    let isolation_active =
+        state.isolated || state.firewall_snapshot.is_some() || state.network_snapshot.is_some();
+    let isolation_arming = isolation_active
+        && state
+            .isolation_started_unix
+            .is_some_and(|started| now.saturating_sub(started) < ISOLATION_ACTIVATION_GRACE_SECS);
+    let probe_reachable = outbound_probe_reachable();
+    let controls_active = isolation_controls_active_best_effort(&state, probe_reachable);
+    if isolation_active && !isolation_arming && !controls_active && probe_reachable {
+        state.isolated = false;
+        state.firewall_snapshot = None;
+        state.network_snapshot = None;
+        state.isolation_started_unix = None;
+        state.isolation_expires_unix = None;
+        if let Err(err) = save_state(&state) {
+            note_state_load_error_once("reconcile_stale_isolation_state_save", &err);
+        }
+        match crate::config::Config::load() {
+            Ok(cfg) => {
+                let _ = crate::break_glass::sync_watchdog(&cfg);
+            }
+            Err(err) => note_state_load_error_once("reconcile_config", &err),
+        }
+        audit::record(
+            "reconcile_stale_isolation_state",
+            "success",
+            json!({
+                "reason": "local isolation controls are no longer active and connectivity probe is reachable"
+            }),
+        );
+        return;
+    }
+    let isolation_expired = isolation_active
+        && state
+            .isolation_expires_unix
+            .is_some_and(|expires_at| now >= expires_at);
+    if isolation_expired {
+        if let Err(err) = restore_machine() {
+            audit::record(
+                "reconcile_isolation_timeout",
+                "error",
+                json!({
+                    "error": err,
+                    "expires_at_unix": state.isolation_expires_unix,
+                    "now_unix": now,
+                }),
+            );
+        } else if let Ok(restored_state) = load_state() {
+            state = restored_state;
+        }
+    }
+    let changed = reconcile_state(&mut state, now, |rule_name| {
+        backend().delete_rule(rule_name).is_ok()
+    });
+    if changed {
+        if let Err(err) = save_state(&state) {
+            note_state_load_error_once("reconcile_save", &err);
+        }
+    }
+}
+
+/// Phase 19 P2: startup-only firewall rule reconciliation.
+///
+/// Re-applies IP-block and isolation rules that may have been lost on
+/// reboot (Linux nftables/iptables are in-memory). Windows WFP filters
+/// persist natively, so this is a no-op there.
+///
+/// Must be called EXACTLY ONCE at boot, NOT from the per-second
+/// reconcile() loop. Uses add-only (never deletes first) so a failed
+/// add does not leave a gap in protection.
+pub fn reconcile_firewall_rules_once() {
+    reconcile_firewall_rules();
+}
+pub fn block_remote(target: &str, preset: DurationPreset) -> Result<String, String> {
+    ensure_modifiable()?;
+    let target = normalise_target(target)?;
+    let rule_name = rule_name_for_target(&target);
+    let expires_at_unix = preset
+        .ttl()
+        .map(|ttl| unix_now().saturating_add(ttl.as_secs()));
+    let _ = backend().delete_rule(&rule_name);
+    backend().add_block_rule(&rule_name, &target)?;
+    let mut state = load_state()?;
+    state.blocked.retain(|rule| rule.target != target);
+    state.blocked.push(BlockedTarget {
+        target: target.clone(),
+        rule_name: rule_name.clone(),
+        expires_at_unix,
+    });
+    save_state(&state)?;
+    let message = match preset {
+        DurationPreset::OneHour => format!("Blocked {target} for 1 hour."),
+        DurationPreset::OneDay => format!("Blocked {target} for 24 hours."),
+        DurationPreset::Permanent => format!("Blocked {target} until removed."),
+    };
+    crate::notifier::send_firewall_event("Rule Added", &message);
+    audit::record(
+        "block_remote",
+        "success",
+        json!({ "target": target, "duration": format!("{:?}", preset), "rule_name": rule_name }),
+    );
+    Ok(message)
+}
+pub fn unblock_remote(target: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    let target = normalise_target(target)?;
+    let mut state = load_state()?;
+    let mut removed = 0usize;
+    let mut kept = Vec::with_capacity(state.blocked.len());
+    let mut delete_failed = false;
+    for rule in state.blocked.drain(..) {
+        if rule.target == target {
+            if backend().delete_rule(&rule.rule_name).is_ok() {
+                removed += 1;
+            } else {
+                delete_failed = true;
+                kept.push(rule);
+            }
+        } else {
+            kept.push(rule);
+        }
+    }
+    state.blocked = kept;
+    if delete_failed {
+        return Err(format!("Could not remove the firewall rule for {target}; it was kept in state so Vigil can retry."));
+    }
+    let message = if removed > 0 {
+        save_state(&state)?;
+        format!("Removed {removed} block rule(s) for {target}.")
+    } else {
+        format!("No active block rule found for {target}.")
+    };
+    audit::record(
+        "unblock_remote",
+        if removed > 0 { "success" } else { "noop" },
+        json!({ "target": target, "removed_rules": removed }),
+    );
+    Ok(message)
+}
+pub fn block_process(pid: u32, path: &str, preset: DurationPreset) -> Result<String, String> {
+    ensure_modifiable()?;
+    let path = normalise_target(path)?;
+    let rule_suffix = rule_suffix_for_process(&path);
+    let outbound_rule_name = format!("{PROCESS_BLOCK_RULE_PREFIX} Out {rule_suffix}");
+    let inbound_rule_name = format!("{PROCESS_BLOCK_RULE_PREFIX} In {rule_suffix}");
+    let expires_at_unix = preset
+        .ttl()
+        .map(|ttl| unix_now().saturating_add(ttl.as_secs()));
+    let _ = backend().delete_rule(&outbound_rule_name);
+    let _ = backend().delete_rule(&inbound_rule_name);
+    backend().add_block_program_rule(&outbound_rule_name, pid, &path, "out")?;
+    if let Err(err) = backend().add_block_program_rule(&inbound_rule_name, pid, &path, "in") {
+        let _ = backend().delete_rule(&outbound_rule_name);
+        return Err(err);
+    }
+    let mut state = load_state()?;
+    state
+        .blocked_processes
+        .retain(|rule| !process_block_matches(rule, &path));
+    state.blocked_processes.push(BlockedProcess {
+        pid,
+        path: path.clone(),
+        inbound_rule_name: inbound_rule_name.clone(),
+        outbound_rule_name: outbound_rule_name.clone(),
+        expires_at_unix,
+    });
+    save_state(&state)?;
+    let message = match preset {
+        DurationPreset::OneHour => format!("Blocked {path} for 1 hour."),
+        DurationPreset::OneDay => format!("Blocked {path} for 24 hours."),
+        DurationPreset::Permanent => format!("Blocked {path} until removed."),
+    };
+    crate::notifier::send_firewall_event("Process Blocked", &message);
+    audit::record(
+        "block_process",
+        "success",
+        json!({ "pid": pid, "path": path, "duration": format!("{:?}", preset), "inbound_rule_name": inbound_rule_name, "outbound_rule_name": outbound_rule_name }),
+    );
+    Ok(message)
+}
+pub fn unblock_process(pid: u32, path: &str) -> Result<String, String> {
+    ensure_modifiable()?;
+    let path = normalise_target(path)?;
+    let mut state = load_state()?;
+    let mut removed = 0usize;
+    let mut kept = Vec::with_capacity(state.blocked_processes.len());
+    let mut delete_failed = false;
+    for rule in state.blocked_processes.drain(..) {
+        if process_block_matches(&rule, &path) {
+            let inbound_deleted = backend().delete_rule(&rule.inbound_rule_name).is_ok();
+            let outbound_deleted = backend().delete_rule(&rule.outbound_rule_name).is_ok();
+            if inbound_deleted && outbound_deleted {
+                removed += 1;
+            } else {
+                delete_failed = true;
+                kept.push(rule);
+            }
+        } else {
+            kept.push(rule);
+        }
+    }
+    state.blocked_processes = kept;
+    if delete_failed {
+        return Err(format!(
+            "Could not remove all firewall rules for {path}; they were kept in state so Vigil can retry."
+        ));
+    }
+    let message = if removed > 0 {
+        save_state(&state)?;
+        format!("Removed {removed} process block(s) for {path}.")
+    } else {
+        format!("No active process block found for {path}.")
+    };
+    audit::record(
+        "unblock_process",
+        if removed > 0 { "success" } else { "noop" },
+        json!({ "pid": pid, "path": path, "removed_entries": removed }),
+    );
+    Ok(message)
+}
+
+pub fn isolate_machine() -> Result<String, String> {
+    ensure_isolation_modifiable()?;
+    let timeout_secs = crate::config::Config::load()
+        .map(|cfg| isolation_timeout_secs(cfg.break_glass_timeout_mins))
+        .unwrap_or(300);
+    let mut state = load_state()?;
+    let firewall_snapshot = platform::snapshot_firewall()?;
+    let network_snapshot = platform::snapshot_active_adapters()?;
+    let previous_state = state.clone();
+    let now = unix_now();
+    stage_isolation_state(
+        &mut state,
+        firewall_snapshot.clone(),
+        network_snapshot.clone(),
+        now,
+        timeout_secs,
+    );
+    save_state(&state)?;
+
+    let mut warnings = Vec::new();
+    let firewall_ok = match platform::outbound_block_supported() {
+        Some(true) => {
+            if let Err(err) = platform::apply_firewall_isolation() {
+                return Err(with_isolation_rollback(
+                    err,
+                    &previous_state,
+                    &firewall_snapshot,
+                ));
+            }
+            true
+        }
+        Some(false) => {
+            if let Err(err) = platform::apply_firewall_isolation_inbound_only() {
+                return Err(with_isolation_rollback(
+                    err,
+                    &previous_state,
+                    &firewall_snapshot,
+                ));
+            }
+            warnings.push(
+                "firewall backend cannot confirm outbound connectivity; using failsafe adapter cutoff"
+                    .to_string(),
+            );
+            false
+        }
+        None => {
+            if let Err(err) = platform::apply_firewall_isolation() {
+                return Err(with_isolation_rollback(
+                    err,
+                    &previous_state,
+                    &firewall_snapshot,
+                ));
+            }
+            true
+        }
+    };
+
+    if !firewall_ok {
+        if let Err(err) = platform::disable_active_adapters(&network_snapshot) {
+            return Err(with_isolation_rollback(
+                err,
+                &previous_state,
+                &firewall_snapshot,
+            ));
+        }
+    }
+
+    match crate::config::Config::load() {
+        Ok(cfg) => {
+            if let Err(err) = crate::break_glass::sync_watchdog(&cfg) {
+                warnings.push(format!("watchdog arm failed: {err}"));
+            }
+        }
+        Err(err) => warnings.push(format!("watchdog configuration load failed: {err}")),
+    }
+
+    crate::notifier::send_firewall_event(
+        "Network Isolated",
+        "Machine network access has been restricted by Vigil.",
+    );
+    audit::record(
+        "isolate_machine",
+        if warnings.is_empty() {
+            "success"
+        } else {
+            "partial"
+        },
+        json!({
+            "warnings": warnings,
+            "timeout_secs": timeout_secs,
+            "mode": if firewall_ok { "firewall" } else { "firewall+adapter_cutoff" }
+        }),
+    );
+
+    if warnings.is_empty() {
+        if firewall_ok {
+            Ok(format!(
+                "Network isolation enabled (failsafe recovery armed: {} minute{} stale-heartbeat timeout)",
+                timeout_secs / 60,
+                if timeout_secs / 60 == 1 { "" } else { "s" }
+            ))
+        } else {
+            Ok(format!(
+                "Network isolation enabled via emergency adapter cutoff (failsafe recovery armed: {} minute{} stale-heartbeat timeout)",
+                timeout_secs / 60,
+                if timeout_secs / 60 == 1 { "" } else { "s" }
+            ))
+        }
+    } else if firewall_ok {
+        Ok(format!(
+            "Network isolation enabled with warnings: {} (failsafe recovery armed: {} minute{} stale-heartbeat timeout)",
+            warnings.join("; "),
+            timeout_secs / 60,
+            if timeout_secs / 60 == 1 { "" } else { "s" }
+        ))
+    } else {
+        Ok(format!(
+            "Network isolation enabled via emergency adapter cutoff with warnings: {} (failsafe recovery armed: {} minute{} stale-heartbeat timeout)",
+            warnings.join("; "),
+            timeout_secs / 60,
+            if timeout_secs / 60 == 1 { "" } else { "s" }
+        ))
+    }
+}
+pub fn restore_machine() -> Result<String, String> {
+    ensure_isolation_modifiable()?;
+    let (mut state, state_load_warning) = match load_state() {
+        Ok(state) => (state, None),
+        Err(err) => {
+            note_state_load_error_once("restore_machine", &err);
+            let state = State {
+                isolated: true,
+                ..State::default()
+            };
+            (state, Some(err))
+        }
+    };
+    let had_isolation_intent = state_load_warning.is_some()
+        || state.isolated
+        || state.firewall_snapshot.is_some()
+        || state.network_snapshot.is_some();
+    let firewall_snapshot = state.firewall_snapshot.clone();
+    let network_snapshot = state.network_snapshot.clone();
+    let mut warnings = Vec::new();
+    if let Some(ref err) = state_load_warning {
+        warnings.push(format!(
+            "protected isolation state could not be loaded; continuing with emergency restore path: {err}"
+        ));
+    }
+    let mut critical_failure = false;
+    let in_deleted = backend().delete_rule(ISOLATE_RULE_IN).is_ok();
+    let out_deleted = backend().delete_rule(ISOLATE_RULE_OUT).is_ok();
+    if !(in_deleted || out_deleted) {
+        if let Some(snapshot) = firewall_snapshot.as_ref() {
+            if let Err(err) = platform::restore_firewall(snapshot) {
+                warnings.push(format!("firewall restore failed: {err}"));
+                critical_failure = true;
+            }
+        }
+    }
+    if let Some(snapshot) = network_snapshot.as_ref() {
+        if let Err(err) = platform::enable_active_adapters(snapshot) {
+            warnings.push(format!(
+                "adapter restore failed: {err}; trying broad adapter re-enable"
+            ));
+            match platform::enable_all_network_adapters() {
+                Ok(enabled) if enabled > 0 => warnings.push(format!(
+                    "broad adapter re-enable restored {enabled} adapter(s)"
+                )),
+                Ok(_) => warnings
+                    .push("broad adapter re-enable did not find any adapters to restore".into()),
+                Err(fallback_err) => {
+                    warnings.push(format!("broad adapter re-enable failed: {fallback_err}"));
+                    critical_failure = true;
+                }
+            }
+        }
+    } else if state_load_warning.is_some() {
+        match platform::enable_all_network_adapters() {
+            Ok(enabled) if enabled > 0 => warnings.push(format!(
+                "broad adapter re-enable restored {enabled} adapter(s) after state-load failure"
+            )),
+            Ok(_) => warnings.push(
+                "broad adapter re-enable did not find any adapters after state-load failure".into(),
+            ),
+            Err(err) => {
+                warnings.push(format!(
+                    "broad adapter re-enable failed after state-load failure: {err}"
+                ));
+                critical_failure = true;
+            }
+        }
+    }
+    match crate::config::Config::load() {
+        Ok(cfg) => {
+            if let Err(err) = crate::break_glass::sync_watchdog(&cfg) {
+                warnings.push(format!("watchdog disarm failed: {err}"));
+            }
+        }
+        Err(err) => warnings.push(format!("watchdog configuration load failed: {err}")),
+    }
+    let _ = backend().delete_rule(ISOLATE_RULE_IN);
+    let _ = backend().delete_rule(ISOLATE_RULE_OUT);
+    state.isolated = false;
+    state.firewall_snapshot = None;
+    state.network_snapshot = None;
+    state.isolation_started_unix = None;
+    state.isolation_expires_unix = None;
+    if let Err(err) = save_state(&state) {
+        warnings.push(format!("state save failed: {err}"));
+    }
+    audit::record(
+        "restore_machine",
+        if critical_failure {
+            "error"
+        } else if warnings.is_empty() {
+            "success"
+        } else {
+            "partial"
+        },
+        json!({ "warnings": warnings }),
+    );
+    if critical_failure {
+        return Err(if warnings.is_empty() {
+            "Failed to restore networking after isolation; manual intervention is required.".into()
+        } else {
+            format!(
+                "Failed to fully restore networking after isolation: {}",
+                warnings.join("; ")
+            )
+        });
+    }
+    if !had_isolation_intent {
+        return Ok("No network isolation state was present; no restore steps were needed.".into());
+    }
+    if warnings.is_empty() {
+        Ok("Network isolation disabled and prior connectivity controls restored.".into())
+    } else {
+        Ok(format!(
+            "Network isolation disabled with warnings: {}",
+            warnings.join("; ")
+        ))
+    }
+}
+
+pub fn is_blocked(target: &str) -> bool {
+    let Ok(target) = normalise_target(target) else {
+        return false;
+    };
+    load_state_for_query("is_blocked")
+        .map(|state| state.blocked.iter().any(|rule| rule.target == target))
+        .unwrap_or(false)
+}
+
+pub fn is_domain_blocked(domain: &str) -> bool {
+    let Ok(domain) = normalise_domain(domain) else {
+        return false;
+    };
+    load_state_for_query("is_domain_blocked")
+        .map(|state| {
+            state
+                .blocked_domains
+                .iter()
+                .any(|entry| entry.domain == domain)
+        })
+        .unwrap_or(false)
+}
+
+pub fn remote_block_remaining(target: &str) -> Option<Duration> {
+    let target = normalise_target(target).ok()?;
+    let now = unix_now();
+    load_state_for_query("remote_block_remaining")
+        .and_then(|state| {
+            state
+                .blocked
+                .iter()
+                .find(|rule| rule.target == target)
+                .and_then(|rule| rule.expires_at_unix)
+        })
+        .and_then(|expires_at_unix| expires_at_unix.checked_sub(now))
+        .map(Duration::from_secs)
+}
+
+pub fn is_process_blocked(_pid: u32, path: &str) -> bool {
+    let Ok(path) = normalise_target(path) else {
+        return false;
+    };
+    load_state_for_query("is_process_blocked")
+        .map(|state| {
+            state
+                .blocked_processes
+                .iter()
+                .any(|rule| process_block_matches(rule, &path))
+        })
+        .unwrap_or(false)
+}
+
+pub fn process_block_remaining(_pid: u32, path: &str) -> Option<Duration> {
+    let path = normalise_target(path).ok()?;
+    let now = unix_now();
+    load_state_for_query("process_block_remaining")
+        .and_then(|state| {
+            state
+                .blocked_processes
+                .iter()
+                .find(|rule| process_block_matches(rule, &path))
+                .and_then(|rule| rule.expires_at_unix)
+        })
+        .and_then(|expires_at_unix| expires_at_unix.checked_sub(now))
+        .map(Duration::from_secs)
+}
+
+pub fn is_process_suspended(pid: u32, path: &str) -> bool {
+    load_state_for_query("is_process_suspended")
+        .map(|state| {
+            state
+                .suspended_processes
+                .iter()
+                .any(|entry| suspended_process_matches(entry, pid, path))
+        })
+        .unwrap_or(false)
+}
+
+fn ensure_modifiable() -> Result<(), String> {
+    if !platform::is_supported() {
+        return Err("Active response is only implemented on supported operating systems.".into());
+    }
+    if !platform::is_elevated() {
+        return Err("Administrator or root privileges are required for active response.".into());
+    }
+    Ok(())
+}
+fn ensure_isolation_modifiable() -> Result<(), String> {
+    if !platform::supports_isolation() {
+        return Err(
+            "Machine isolation is not available on this operating system or backend yet.".into(),
+        );
+    }
+    if !platform::is_elevated() {
+        return Err("Administrator or root privileges are required for network isolation.".into());
+    }
+    Ok(())
+}
+fn normalise_target(target: &str) -> Result<String, String> {
+    let trimmed = target.trim();
+    if trimmed.is_empty() {
+        return Err("Target cannot be empty.".into());
+    }
+    Ok(trimmed.to_string())
+}
+fn normalise_domain(domain: &str) -> Result<String, String> {
+    let trimmed = domain.trim().trim_end_matches('.').to_ascii_lowercase();
+    if trimmed.is_empty() {
+        return Err("Domain cannot be empty.".into());
+    }
+    if trimmed.parse::<IpAddr>().is_ok() {
+        return Err("Domain blocking expects a hostname, not an IP address.".into());
+    }
+    if !trimmed
+        .chars()
+        .all(|ch| ch.is_ascii_alphanumeric() || matches!(ch, '.' | '-'))
+    {
+        return Err("Domain contains unsupported characters.".into());
+    }
+    if !trimmed.contains('.') {
+        return Err("Domain must contain at least one dot.".into());
+    }
+    Ok(trimmed)
+}
+fn rule_name_for_target(target: &str) -> String {
+    format!("{BLOCK_RULE_PREFIX} {target}")
+}
+fn rule_suffix_for_process(path: &str) -> String {
+    let normalized_path = path.to_ascii_lowercase();
+    let readable_prefix: String = normalized_path
+        .chars()
+        .map(|ch| if ch.is_ascii_alphanumeric() { ch } else { '_' })
+        .take(PROCESS_RULE_PREFIX_LEN)
+        .collect();
+    let readable_prefix = readable_prefix.trim_matches('_');
+    let readable_prefix = if readable_prefix.is_empty() {
+        "process"
+    } else {
+        readable_prefix
+    };
+    let fingerprint = hex_prefix(
+        &Sha256::digest(normalized_path.as_bytes()),
+        PROCESS_RULE_FINGERPRINT_LEN,
+    );
+    format!("{readable_prefix}_{fingerprint}")
+}
+fn domain_marker(domain: &str) -> String {
+    format!("{DOMAIN_MARKER_PREFIX} {domain}")
+}
+fn reconcile_state<F>(state: &mut State, now: u64, mut delete_rule: F) -> bool
+where
+    F: FnMut(&str) -> bool,
+{
+    let mut changed = false;
+    let mut kept = Vec::with_capacity(state.blocked.len());
+    for rule in state.blocked.drain(..) {
+        let expired = rule.expires_at_unix.is_some_and(|deadline| deadline <= now);
+        if expired && delete_rule(&rule.rule_name) {
+            changed = true;
+            continue;
+        }
+        kept.push(rule);
+    }
+    state.blocked = kept;
+
+    let mut kept_processes = Vec::with_capacity(state.blocked_processes.len());
+    for rule in state.blocked_processes.drain(..) {
+        let expired = rule.expires_at_unix.is_some_and(|deadline| deadline <= now);
+        if expired {
+            let inbound_deleted = delete_rule(&rule.inbound_rule_name);
+            let outbound_deleted = delete_rule(&rule.outbound_rule_name);
+            if inbound_deleted && outbound_deleted {
+                changed = true;
+                continue;
+            }
+        }
+        kept_processes.push(rule);
+    }
+    state.blocked_processes = kept_processes;
+    changed
+}
+fn process_block_matches(rule: &BlockedProcess, path: &str) -> bool {
+    rule.path.eq_ignore_ascii_case(path)
+}
+fn suspended_process_matches(entry: &SuspendedProcess, pid: u32, path: &str) -> bool {
+    entry.pid == pid && (path.is_empty() || entry.path.eq_ignore_ascii_case(path))
+}
+pub fn extract_remote_target(addr: &str) -> Option<String> {
+    addr.rsplit_once(':').and_then(|(host, _)| {
+        if host.starts_with('[') && host.ends_with(']') {
+            Some(host.trim_matches(&['[', ']'][..]).to_string())
+        } else if host.contains(':') {
+            Some(host.to_string())
+        } else {
+            host.parse::<IpAddr>().ok().map(|_| host.to_string())
+        }
+    })
+}
+pub fn extract_domain_target(conn: &ConnInfo) -> Option<String> {
+    conn.hostname
+        .as_deref()
+        .and_then(extract_domain_from_hostname)
+}
+fn extract_domain_from_hostname(hostname: &str) -> Option<String> {
+    let trimmed = hostname.trim().trim_end_matches('.');
+    if trimmed.is_empty() || trimmed.parse::<IpAddr>().is_ok() {
+        None
+    } else {
+        Some(trimmed.to_ascii_lowercase())
+    }
+}
+fn parse_socket_addr(addr: &str, label: &str) -> Result<SocketAddr, SocketKillError> {
+    addr.parse::<SocketAddr>().map_err(|_| match label {
+        "local" => SocketKillError::InvalidLocalAddr(addr.into()),
+        _ => SocketKillError::InvalidRemoteAddr(addr.into()),
+    })
+}
+#[allow(dead_code)]
+fn socket_addr_from_text(text: &str) -> Result<SocketAddr, std::net::AddrParseError> {
+    text.trim().parse::<SocketAddr>()
+}
+fn socket_kill_target(conn: &ConnInfo) -> Result<SocketKillTarget, SocketKillError> {
+    let status = conn.status.trim();
+    if status != "ESTABLISHED" {
+        return Err(SocketKillError::UnsupportedStatus(status.into()));
+    }
+    let local = parse_socket_addr(&conn.local_addr, "local")?;
+    let remote = parse_socket_addr(&conn.remote_addr, "remote")?;
+    if !matches!(local.ip(), IpAddr::V4(_)) || !matches!(remote.ip(), IpAddr::V4(_)) {
+        return Err(SocketKillError::UnsupportedAddressFamily);
+    }
+    Ok(SocketKillTarget { local, remote })
+}
+fn unix_now() -> u64 {
+    SystemTime::now()
+        .duration_since(UNIX_EPOCH)
+        .unwrap_or_default()
+        .as_secs()
+}
+
+fn isolation_timeout_secs(timeout_mins: u64) -> u64 {
+    timeout_mins.clamp(1, ISOLATION_MAX_SECS / 60) * 60
+}
+
+fn stage_isolation_state(
+    state: &mut State,
+    firewall_snapshot: FirewallSnapshot,
+    network_snapshot: NetworkSnapshot,
+    now: u64,
+    timeout_secs: u64,
+) {
+    state.firewall_snapshot = Some(firewall_snapshot);
+    state.network_snapshot = Some(network_snapshot);
+    state.isolated = true;
+    state.isolation_started_unix = Some(now);
+    state.isolation_expires_unix = Some(now.saturating_add(timeout_secs));
+}
+
+fn rollback_isolation_state(
+    previous_state: &State,
+    firewall_snapshot: &FirewallSnapshot,
+) -> Result<(), String> {
+    let mut rollback_errors = Vec::new();
+    if let Err(err) = platform::restore_firewall(firewall_snapshot) {
+        rollback_errors.push(format!("restore firewall: {err}"));
+    }
+    if let Err(err) = save_state(previous_state) {
+        rollback_errors.push(format!("restore saved state: {err}"));
+    }
+    if rollback_errors.is_empty() {
+        Ok(())
+    } else {
+        Err(format!(
+            "failed to roll back partial isolation attempt: {}",
+            rollback_errors.join("; ")
+        ))
+    }
+}
+
+fn with_isolation_rollback(
+    err: String,
+    previous_state: &State,
+    firewall_snapshot: &FirewallSnapshot,
+) -> String {
+    match rollback_isolation_state(previous_state, firewall_snapshot) {
+        Ok(()) => err,
+        Err(rollback_err) => format!("{err}; {rollback_err}"),
+    }
+}
+
+fn outbound_probe_reachable() -> bool {
+    TcpStream::connect_timeout(
+        &SocketAddr::from(([1, 1, 1, 1], 443)),
+        Duration::from_secs(2),
+    )
+    .is_ok()
+}
+#[allow(dead_code)]
+fn wait_for_outbound_probe(timeout: Duration) -> bool {
+    let start = std::time::Instant::now();
+    loop {
+        if outbound_probe_reachable() {
+            return true;
+        }
+        if start.elapsed() >= timeout {
+            return false;
+        }
+        std::thread::sleep(Duration::from_millis(650));
+    }
+}
+fn isolation_effective_now(state: &State, now: u64) -> bool {
+    if state
+        .isolation_expires_unix
+        .is_some_and(|deadline| deadline <= now)
+    {
+        return false;
+    }
+    if !state.isolated && state.firewall_snapshot.is_none() && state.network_snapshot.is_none() {
+        return false;
+    }
+    let probe_reachable = state.network_snapshot.is_some() && outbound_probe_reachable();
+    isolation_controls_active_best_effort(state, probe_reachable)
+}
+fn isolation_controls_active_best_effort(state: &State, probe_reachable: bool) -> bool {
+    let controls_active = match platform::isolation_controls_active(state) {
+        Ok(active) => active,
+        Err(_) => state.firewall_snapshot.is_some() || state.network_snapshot.is_some(),
+    };
+    isolation_controls_still_effective(state, controls_active, probe_reachable)
+}
+fn isolation_controls_still_effective(
+    state: &State,
+    controls_active: bool,
+    probe_reachable: bool,
+) -> bool {
+    if !controls_active {
+        return false;
+    }
+    if state.network_snapshot.is_some() && probe_reachable {
+        return false;
+    }
+    true
+}
+
+fn hex_prefix(bytes: &[u8], hex_len: usize) -> String {
+    let mut out = String::with_capacity(hex_len);
+    for byte in bytes.iter().take(hex_len.div_ceil(2)) {
+        let hi = byte >> 4;
+        let lo = byte & 0x0f;
+        out.push(char::from(b"0123456789abcdef"[hi as usize]));
+        if out.len() == hex_len {
+            break;
+        }
+        out.push(char::from(b"0123456789abcdef"[lo as usize]));
+        if out.len() == hex_len {
+            break;
+        }
+    }
+    out
+}
+
+static STATE_LOAD_WARNING_ONCE: OnceLock<Mutex<std::collections::HashSet<String>>> =
+    OnceLock::new();
+fn note_state_load_error_once(context: &str, err: &str) {
+    let seen = STATE_LOAD_WARNING_ONCE.get_or_init(|| Mutex::new(std::collections::HashSet::new()));
+    let key = format!("{context}:{err}");
+    let mut seen = seen.lock().unwrap();
+    if seen.insert(key) {
+        tracing::warn!(context, %err, "active-response state load failed");
+    }
+}
+
+static STATE_PATH_OVERRIDE: OnceLock<RwLock<Option<PathBuf>>> = OnceLock::new();
+#[derive(Clone)]
+struct QueryStateCache {
+    path: PathBuf,
+    signature: Option<StateFileSignature>,
+    loaded_at: Instant,
+    state: State,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+struct StateFileSignature {
+    modified: Option<SystemTime>,
+    len: u64,
+}
+
+static QUERY_STATE_CACHE: OnceLock<RwLock<Option<QueryStateCache>>> = OnceLock::new();
+
+fn state_path() -> PathBuf {
+    let override_lock = STATE_PATH_OVERRIDE.get_or_init(|| RwLock::new(None));
+    if let Some(path) = override_lock.read().unwrap().clone() {
+        return path;
+    }
+    crate::config::data_dir().join(STATE_FILE)
+}
+fn load_state() -> Result<State, String> {
+    let path = state_path();
+    load_state_from_path(&path)
+}
+fn load_state_for_query(context: &str) -> Option<State> {
+    let path = state_path();
+    let cache_lock = QUERY_STATE_CACHE.get_or_init(|| RwLock::new(None));
+    {
+        let cache = cache_lock.read().unwrap();
+        if let Some(cache) = cache.as_ref() {
+            if cache.path == path && cache.loaded_at.elapsed() <= QUERY_STATE_CACHE_TTL {
+                return Some(cache.state.clone());
+            }
+        }
+    }
+
+    let signature = state_file_signature(&path);
+    {
+        let mut cache = cache_lock.write().unwrap();
+        if let Some(existing) = cache.as_mut() {
+            if existing.path == path && existing.signature == signature {
+                existing.loaded_at = Instant::now();
+                return Some(existing.state.clone());
+            }
+        }
+    }
+
+    match load_state_from_path(&path) {
+        Ok(state) => Some(state),
+        Err(err) => {
+            note_state_load_error_once(context, &err);
+            None
+        }
+    }
+    .inspect(|state| {
+        let mut cache = cache_lock.write().unwrap();
+        *cache = Some(QueryStateCache {
+            path,
+            signature,
+            loaded_at: Instant::now(),
+            state: state.clone(),
+        });
+    })
+}
+
+fn state_file_signature(path: &Path) -> Option<StateFileSignature> {
+    path.metadata().ok().map(|metadata| StateFileSignature {
+        modified: metadata.modified().ok(),
+        len: metadata.len(),
+    })
+}
+
+fn update_query_state_cache(path: &Path, state: &State) {
+    let cache_lock = QUERY_STATE_CACHE.get_or_init(|| RwLock::new(None));
+    let mut cache = cache_lock.write().unwrap();
+    *cache = Some(QueryStateCache {
+        path: path.to_path_buf(),
+        signature: state_file_signature(path),
+        loaded_at: Instant::now(),
+        state: state.clone(),
+    });
+}
+
+fn clear_query_state_cache() {
+    let cache_lock = QUERY_STATE_CACHE.get_or_init(|| RwLock::new(None));
+    let mut cache = cache_lock.write().unwrap();
+    *cache = None;
+}
+
+fn save_state(state: &State) -> Result<(), String> {
+    let data = serde_json::to_vec_pretty(state)
+        .map_err(|e| format!("serialize active response state: {e}"))?;
+    let path = state_path();
+    match crate::security::policy::save_json_with_integrity(&path, &data) {
+        Ok(()) => {
+            update_query_state_cache(&path, state);
+            sync_firewall_rules_to_db(state);
+            Ok(())
+        }
+        Err(err) => {
+            clear_query_state_cache();
+            Err(err)
+        }
+    }
+}
+
+fn sync_firewall_rules_to_db(state: &State) {
+    let rows: Vec<crate::storage::db::FirewallRuleRow> = state
+        .blocked
+        .iter()
+        .map(|b| crate::storage::db::FirewallRuleRow {
+            rule_name: b.rule_name.clone(),
+            rule_type: "ip".into(),
+            target: b.target.clone(),
+            direction: "out".into(),
+            pid: 0,
+            path: String::new(),
+            created_unix: 0,
+            expires_unix: b.expires_at_unix,
+        })
+        .chain(state.blocked_processes.iter().flat_map(|b| {
+            let out = crate::storage::db::FirewallRuleRow {
+                rule_name: b.outbound_rule_name.clone(),
+                rule_type: "process".into(),
+                target: String::new(),
+                direction: "out".into(),
+                pid: b.pid,
+                path: b.path.clone(),
+                created_unix: 0,
+                expires_unix: b.expires_at_unix,
+            };
+            let in_rule = crate::storage::db::FirewallRuleRow {
+                rule_name: b.inbound_rule_name.clone(),
+                rule_type: "process".into(),
+                target: String::new(),
+                direction: "in".into(),
+                pid: b.pid,
+                path: b.path.clone(),
+                created_unix: 0,
+                expires_unix: b.expires_at_unix,
+            };
+            [out, in_rule]
+        }))
+        .chain(
+            state
+                .blocked_domains
+                .iter()
+                .map(|d| crate::storage::db::FirewallRuleRow {
+                    rule_name: format!("domain-{}", d.domain),
+                    rule_type: "domain".into(),
+                    target: d.domain.clone(),
+                    direction: "out".into(),
+                    pid: 0,
+                    path: String::new(),
+                    created_unix: 0,
+                    expires_unix: None,
+                }),
+        )
+        .collect();
+    if let Ok(db) = crate::storage::db::StorageDb::global() {
+        if let Err(e) = db.save_firewall_rules(&rows) {
+            tracing::warn!("sync firewall rules to db: {e}");
+        }
+        if let Err(e) = db.checkpoint() {
+            tracing::warn!("checkpoint after firewall sync: {e}");
+        }
+    }
+}
+
+fn load_state_from_path(path: &std::path::Path) -> Result<State, String> {
+    let existed = path.exists();
+    match crate::security::policy::load_struct_with_integrity::<State>(path) {
+        Ok(Some(state)) => Ok(state),
+        Ok(None) => {
+            if existed {
+                Err(format!(
+                    "active-response state {} exists but could not be verified or restored",
+                    path.display()
+                ))
+            } else {
+                Ok(State::default())
+            }
+        }
+        Err(err) => Err(err),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn blocked_target(target: &str, expires_at_unix: Option<u64>) -> BlockedTarget {
+        BlockedTarget {
+            target: target.to_string(),
+            rule_name: format!("rule-{target}"),
+            expires_at_unix,
+        }
+    }
+    fn blocked_process(path: &str, expires_at_unix: Option<u64>) -> BlockedProcess {
+        BlockedProcess {
+            pid: 1234,
+            path: path.to_string(),
+            inbound_rule_name: format!("in-{path}"),
+            outbound_rule_name: format!("out-{path}"),
+            expires_at_unix,
+        }
+    }
+    fn blocked_domain(domain: &str) -> BlockedDomain {
+        BlockedDomain {
+            domain: domain.to_string(),
+            marker: domain_marker(domain),
+        }
+    }
+    fn suspended_process(pid: u32, path: &str) -> SuspendedProcess {
+        SuspendedProcess {
+            pid,
+            path: path.to_string(),
+            proc_name: "app.exe".into(),
+            suspended_at_unix: 10,
+        }
+    }
+    fn conn(local: &str, remote: &str, status: &str) -> ConnInfo {
+        ConnInfo {
+            timestamp: "12:00:00".into(),
+            proc_name: "evil.exe".into(),
+            pid: 4242,
+            proc_path: "C:/Temp/evil.exe".into(),
+            proc_user: "user".into(),
+            parent_name: "cmd.exe".into(),
+            parent_pid: 123,
+            parent_user: "user".into(),
+            service_name: String::new(),
+            publisher: String::new(),
+            local_addr: local.into(),
+            remote_addr: remote.into(),
+            status: status.into(),
+            protocol: crate::types::TransportProtocol::Tcp,
+            score: 9,
+            reasons: vec!["test".into()],
+            ancestor_chain: vec![("cmd.exe".into(), 123)],
+            pre_login: false,
+            hostname: Some("c2.bad.example".into()),
+            country: None,
+            asn: None,
+            asn_org: None,
+            reputation_hit: None,
+            recently_dropped: false,
+            long_lived: false,
+            dga_like: false,
+            baseline_deviation: false,
+            script_host_suspicious: false,
+            command_line: String::new(),
+            attack_tags: Vec::new(),
+            tls_sni: None,
+            tls_ja3: None,
+        }
+    }
+    #[test]
+    fn reconcile_keeps_expired_entries_when_deletion_fails() {
+        let mut state = State {
+            blocked: vec![blocked_target("10.0.0.1", Some(10))],
+            blocked_processes: vec![blocked_process("C:/app.exe", Some(10))],
+            blocked_domains: vec![blocked_domain("bad.example")],
+            suspended_processes: vec![],
+            autorun_snapshot: None,
+            firewall_snapshot: None,
+            network_snapshot: None,
+            isolated: false,
+            isolation_started_unix: None,
+            isolation_expires_unix: None,
+        };
+        let changed = reconcile_state(&mut state, 100, |_| false);
+        assert!(!changed);
+        assert_eq!(state.blocked.len(), 1);
+        assert_eq!(state.blocked_processes.len(), 1);
+        assert_eq!(state.blocked_domains.len(), 1);
+    }
+    #[test]
+    fn reconcile_removes_expired_entries_after_successful_deletion() {
+        let mut deleted = Vec::new();
+        let mut state = State {
+            blocked: vec![blocked_target("10.0.0.1", Some(10))],
+            blocked_processes: vec![blocked_process("C:/app.exe", Some(10))],
+            blocked_domains: vec![],
+            suspended_processes: vec![],
+            autorun_snapshot: None,
+            firewall_snapshot: None,
+            network_snapshot: None,
+            isolated: false,
+            isolation_started_unix: None,
+            isolation_expires_unix: None,
+        };
+        let changed = reconcile_state(&mut state, 100, |rule| {
+            deleted.push(rule.to_string());
+            true
+        });
+        assert!(changed);
+        assert!(state.blocked.is_empty());
+        assert!(state.blocked_processes.is_empty());
+        assert_eq!(
+            deleted,
+            vec![
+                "rule-10.0.0.1".to_string(),
+                "in-C:/app.exe".to_string(),
+                "out-C:/app.exe".to_string()
+            ]
+        );
+    }
+    #[test]
+    fn process_blocks_match_by_path_only() {
+        let rule = BlockedProcess {
+            pid: 1234,
+            path: "C:/app.exe".into(),
+            inbound_rule_name: "in".into(),
+            outbound_rule_name: "out".into(),
+            expires_at_unix: Some(200),
+        };
+        assert!(process_block_matches(&rule, "C:/app.exe"));
+        assert!(!process_block_matches(&rule, "C:/other.exe"));
+    }
+    #[test]
+    fn suspended_processes_match_on_pid_and_optional_path() {
+        let entry = suspended_process(1234, "C:/app.exe");
+        assert!(suspended_process_matches(&entry, 1234, "C:/app.exe"));
+        assert!(suspended_process_matches(&entry, 1234, ""));
+        assert!(!suspended_process_matches(&entry, 9999, "C:/app.exe"));
+    }
+    #[test]
+    fn socket_kill_target_accepts_established_ipv4_connections() {
+        let parsed =
+            socket_kill_target(&conn("192.168.1.10:50000", "8.8.8.8:443", "ESTABLISHED")).unwrap();
+        assert_eq!(parsed.local.to_string(), "192.168.1.10:50000");
+        assert_eq!(parsed.remote.to_string(), "8.8.8.8:443");
+    }
+    #[test]
+    fn socket_kill_target_rejects_listen_rows() {
+        let err = socket_kill_target(&conn("0.0.0.0:80", "0.0.0.0:0", "LISTEN")).unwrap_err();
+        assert!(matches!(err, SocketKillError::UnsupportedStatus(_)));
+    }
+    #[test]
+    fn socket_kill_target_rejects_invalid_remote() {
+        let err = socket_kill_target(&conn(
+            "192.168.1.10:50000",
+            "not-an-endpoint",
+            "ESTABLISHED",
+        ))
+        .unwrap_err();
+        assert!(matches!(err, SocketKillError::InvalidRemoteAddr(_)));
+    }
+    #[test]
+    fn socket_kill_target_rejects_ipv6_for_now() {
+        let err = socket_kill_target(&conn(
+            "[::1]:50000",
+            "[2606:4700:4700::1111]:443",
+            "ESTABLISHED",
+        ))
+        .unwrap_err();
+        assert!(matches!(err, SocketKillError::UnsupportedAddressFamily));
+    }
+    #[test]
+    fn extract_remote_target_supports_ipv4_and_ipv6() {
+        assert_eq!(
+            extract_remote_target("8.8.8.8:443").as_deref(),
+            Some("8.8.8.8")
+        );
+        assert_eq!(
+            extract_remote_target("[2606:4700:4700::1111]:443").as_deref(),
+            Some("2606:4700:4700::1111")
+        );
+        assert_eq!(
+            extract_remote_target("2606:4700:4700::1111:443").as_deref(),
+            Some("2606:4700:4700::1111")
+        );
+    }
+    #[test]
+    fn extract_domain_target_uses_hostname() {
+        let sample = conn("192.168.1.10:50000", "8.8.8.8:443", "ESTABLISHED");
+        assert_eq!(
+            extract_domain_target(&sample).as_deref(),
+            Some("c2.bad.example")
+        );
+        assert_eq!(extract_domain_from_hostname("8.8.8.8"), None);
+    }
+    #[test]
+    fn process_rule_suffix_keeps_readable_prefix_and_unique_fingerprint() {
+        let left = rule_suffix_for_process("/tmp/a-b");
+        let right = rule_suffix_for_process("/tmp/a_b");
+
+        assert!(left.starts_with("tmp_a_b_"));
+        assert!(right.starts_with("tmp_a_b_"));
+        assert_ne!(left, right);
+    }
+    #[test]
+    fn isolation_controls_guard_uses_live_probe_for_network_snapshots() {
+        let state = State {
+            blocked: vec![],
+            blocked_processes: vec![],
+            blocked_domains: vec![],
+            suspended_processes: vec![],
+            autorun_snapshot: None,
+            firewall_snapshot: None,
+            network_snapshot: Some(NetworkSnapshot {
+                adapters: vec![NetworkAdapterState {
+                    name: "Ethernet".into(),
+                    is_wireless: false,
+                    wifi_profile: None,
+                }],
+            }),
+            isolated: false,
+            isolation_started_unix: None,
+            isolation_expires_unix: None,
+        };
+
+        assert!(isolation_controls_still_effective(&state, true, false));
+        assert!(!isolation_controls_still_effective(&state, true, true));
+        assert!(!isolation_controls_still_effective(&state, false, false));
+    }
+
+    #[test]
+    fn isolation_timeout_preserves_existing_one_hour_cap() {
+        assert_eq!(isolation_timeout_secs(0), 60);
+        assert_eq!(isolation_timeout_secs(10), 600);
+        assert_eq!(isolation_timeout_secs(60), ISOLATION_MAX_SECS);
+        assert_eq!(isolation_timeout_secs(240), ISOLATION_MAX_SECS);
+    }
+
+    #[test]
+    fn staging_isolation_state_sets_timeout_and_snapshots() {
+        let mut state = State::default();
+        let firewall_snapshot = FirewallSnapshot { profiles: vec![] };
+        let network_snapshot = NetworkSnapshot { adapters: vec![] };
+
+        stage_isolation_state(
+            &mut state,
+            firewall_snapshot.clone(),
+            network_snapshot.clone(),
+            100,
+            600,
+        );
+
+        assert!(state.isolated);
+        assert_eq!(state.isolation_started_unix, Some(100));
+        assert_eq!(state.isolation_expires_unix, Some(700));
+        assert_eq!(state.firewall_snapshot, Some(firewall_snapshot));
+        assert_eq!(state.network_snapshot, Some(network_snapshot));
+    }
+
+    #[test]
+    fn normalise_target_trims_whitespace() {
+        assert_eq!(normalise_target(" 10.0.0.1 ").unwrap(), "10.0.0.1");
+    }
+
+    #[test]
+    fn normalise_target_rejects_empty() {
+        assert!(normalise_target("").is_err());
+        assert!(normalise_target("  ").is_err());
+    }
+
+    #[test]
+    fn normalise_domain_trims_lowercases_and_strips_trailing_dot() {
+        assert_eq!(normalise_domain("  Example.COM. ").unwrap(), "example.com");
+    }
+
+    #[test]
+    fn normalise_domain_rejects_ip_address() {
+        assert!(normalise_domain("8.8.8.8").is_err());
+    }
+
+    #[test]
+    fn normalise_domain_rejects_no_dot() {
+        assert!(normalise_domain("localhost").is_err());
+    }
+
+    #[test]
+    fn normalise_domain_rejects_special_chars() {
+        assert!(normalise_domain("bad space.com").is_err());
+    }
+
+    #[test]
+    fn rule_name_for_target_formats_correctly() {
+        assert_eq!(rule_name_for_target("10.0.0.1"), "Vigil Block 10.0.0.1");
+    }
+
+    #[test]
+    fn domain_marker_formats_correctly() {
+        assert_eq!(
+            domain_marker("evil.example"),
+            "# Vigil Domain Block evil.example"
+        );
+    }
+
+    #[test]
+    fn hex_prefix_produces_requested_length() {
+        let bytes = [0xab, 0xcd, 0xef];
+        assert_eq!(hex_prefix(&bytes, 4), "abcd");
+        assert_eq!(hex_prefix(&bytes, 6), "abcdef");
+        assert_eq!(hex_prefix(&bytes, 0), "");
+    }
+
+    #[test]
+    fn hex_prefix_handles_empty_bytes() {
+        assert_eq!(hex_prefix(&[], 4), "");
+    }
+
+    #[test]
+    fn extract_domain_from_hostname_returns_none_for_ip() {
+        assert!(extract_domain_from_hostname("8.8.8.8").is_none());
+    }
+
+    #[test]
+    fn extract_domain_from_hostname_lowercases_and_strips_dot() {
+        assert_eq!(
+            extract_domain_from_hostname("Bad.Example.COM.").unwrap(),
+            "bad.example.com"
+        );
+    }
+
+    #[test]
+    fn extract_domain_from_hostname_returns_none_for_empty() {
+        assert!(extract_domain_from_hostname("").is_none());
+    }
+
+    #[test]
+    fn parse_socket_addr_accepts_valid_addresses() {
+        assert!(parse_socket_addr("192.168.1.1:443", "remote").is_ok());
+        assert!(parse_socket_addr("[::1]:80", "local").is_ok());
+    }
+
+    #[test]
+    fn parse_socket_addr_rejects_port_only() {
+        assert!(parse_socket_addr(":443", "remote").is_err());
+    }
+
+    #[test]
+    fn socket_kill_target_requires_established_status() {
+        let c = conn("10.0.0.1:50000", "8.8.8.8:443", "TIME_WAIT");
+        assert!(matches!(
+            socket_kill_target(&c).unwrap_err(),
+            SocketKillError::UnsupportedStatus(_)
+        ));
+    }
+
+    #[test]
+    fn socket_kill_target_accepts_localhost_remote_as_valid_address() {
+        let parsed =
+            socket_kill_target(&conn("192.168.1.10:50000", "127.0.0.1:443", "ESTABLISHED"))
+                .unwrap();
+        assert_eq!(parsed.remote.to_string(), "127.0.0.1:443");
+    }
+
+    #[test]
+    fn duration_presets_have_distinct_ttls() {
+        use std::time::Duration;
+        assert_eq!(
+            DurationPreset::OneHour.ttl(),
+            Some(Duration::from_secs(3600))
+        );
+        assert_eq!(
+            DurationPreset::OneDay.ttl(),
+            Some(Duration::from_secs(86400))
+        );
+        assert_eq!(DurationPreset::Permanent.ttl(), None);
+    }
+
+    #[test]
+    fn duration_presets_labels_are_non_empty() {
+        assert!(!DurationPreset::OneHour.label().is_empty());
+        assert!(!DurationPreset::OneDay.label().is_empty());
+        assert!(!DurationPreset::Permanent.label().is_empty());
+    }
+
+    #[test]
+    fn reconcile_state_removes_non_expired_entries_unchanged() {
+        let mut state = State {
+            blocked: vec![blocked_target("10.0.0.1", Some(200))],
+            ..State::default()
+        };
+        let changed = reconcile_state(&mut state, 100, |_| unreachable!());
+        assert!(!changed);
+        assert_eq!(state.blocked.len(), 1);
+    }
+
+    #[test]
+    fn reconcile_state_handles_empty_state() {
+        let mut state = State::default();
+        let changed = reconcile_state(&mut state, 100, |_| unreachable!());
+        assert!(!changed);
+    }
+
+    #[test]
+    fn reconcile_state_handles_non_expired_domains() {
+        let mut state = State {
+            blocked_domains: vec![blocked_domain("bad.example")],
+            ..State::default()
+        };
+        let changed = reconcile_state(&mut state, 100, |_| unreachable!());
+        assert!(!changed);
+    }
+
+    #[test]
+    fn process_block_matches_rejects_empty_path() {
+        let rule = BlockedProcess {
+            pid: 1234,
+            path: "C:/app.exe".into(),
+            inbound_rule_name: "in".into(),
+            outbound_rule_name: "out".into(),
+            expires_at_unix: None,
+        };
+        assert!(!process_block_matches(&rule, ""));
+    }
+
+    #[test]
+    fn suspended_process_matches_pid_and_matching_path() {
+        let entry = suspended_process(42, "C:/app.exe");
+        assert!(suspended_process_matches(&entry, 42, "C:/app.exe"));
+        assert!(suspended_process_matches(&entry, 42, ""));
+        assert!(!suspended_process_matches(&entry, 99, "C:/app.exe"));
+        assert!(!suspended_process_matches(&entry, 42, "C:/other.exe"));
+    }
+
+    #[test]
+    fn isolation_controls_still_effective_probe_reachable_relieves() {
+        let state = State {
+            network_snapshot: Some(NetworkSnapshot {
+                adapters: vec![NetworkAdapterState {
+                    name: "Ethernet".into(),
+                    is_wireless: false,
+                    wifi_profile: None,
+                }],
+            }),
+            ..State::default()
+        };
+        assert!(!isolation_controls_still_effective(&state, true, true));
+        assert!(isolation_controls_still_effective(&state, true, false));
+    }
+
+    #[test]
+    fn isolation_controls_still_effective_inactive_controls() {
+        let state = State::default();
+        assert!(!isolation_controls_still_effective(&state, false, false));
+    }
+
+    #[test]
+    fn process_block_rule_suffix_differs_for_equivalent_paths() {
+        let left = rule_suffix_for_process("C:/Temp/evil.exe");
+        let right = rule_suffix_for_process("C:\\Temp\\evil.exe");
+        assert_ne!(left, right);
+    }
+
+    #[test]
+    fn process_block_rule_suffix_starts_with_readable_prefix() {
+        let suffix = rule_suffix_for_process("/usr/bin/suspicious");
+        assert!(suffix.starts_with("usr_bin_suspicious_"));
+    }
+}
